Consumer/requirements.txt
kafka-python
pyspark
cryptography
python-dotenv